Finishing TODOs
- Added maybe_unused to supress warnings from pallete.h
- Organized pallete colors (remove unused, and sort by "importance")
- :mr turn off selection
- utils.h is now a single header for multiple files (all inside src/util/)
- File errors handling on read/write operations
- Removed read_str(uint) from editor*

* The normal read_str method is used to read a fixed
length string, which is ok, but the read_str(uint) was planned to
automatically expand string to the next null byte character. The problem with it
is that the file may not have a null byte. In this case, the entire file will be
loaded as string, which is not a good idea for big files (>1Gb).